Clearer example for "description_from" attribute in Help.
Re-ordered the (comma-separated) values of "description_from" attribute
in a way that makes, imo, a clearer example of its usage.

Changed from :

    Example: @description_from="body,excerpt,metadescription"@  Default: @"body,excerpt"@

To:

    Example: @description_from="metadescription,excerpt,body"@  Default: @"body,excerpt"@

The rationale is: if user has implemented a "metadescription" custom field, it's probably
that user wants the description fetched/generated from that custom field first.
Then, if that field is empty, it will fallback to "excerpt", and, as last resort, to "body".
In other words, it wouldn't make sense, imo, to try to fetch/generate the description string
from the "body" field first, instead of using the content of the "metadescription" field,
if available.

By applying that same reasoning, I'd like to suggest changing the default value ("body,excerpt")
to "excerpt,body". IMO, it's probably a more real use case to re-purpose the "excerpt" field
into a field that holds the meta description (or even just its classic/common usage, as a short
summary that could serve fine also as start point to generate a meta description), than resorting
first to the contents of the "body" field . Considering that is highly probably that the "body"
field  won't be empty, I would use the contents of "body" as a last resort.