Adding a rah_meta_description function (also a txp tag) to re-use rah_metas string processing powers.
This rah_meta_description function/tag will take advantage of rah_metas string
processing powers to generate and output just a (processed) description string.
In other words, the function will not print the full HTML tag
(ie. <meta name="..." />), but just the string (the one that rah_metas would
have feed into the content="" attribute of <meta /> element).

The motivation/rationale for adding this was to be able to use the rah_metas powers
to generate the content for a <meta property="og:description" content="..." />
tag.